reduce test code setup/teardown duplication
